Change DMX Offsets/Breaks to u16
This includes the 9-bit, 512-channels of DMX, but is not bigger
and disallows negative values.

+++ src/errors.rs
@@ -58,7 +58,7 @@ pub enum Problem {
     #[error("duplicate Geometry name '{0}' at line {1}")]
     DuplicateGeometryName(String, TextPos),
     #[error("duplicate DMXBreak attribute '{0}' at line {1}")]
-    DuplicateDmxBreak(u32, TextPos),
+    DuplicateDmxBreak(u16, TextPos),
     #[error("unexpected 'GeometryReference' as top-level Geometry at line {0}")]
     UnexpectedTopLevelGeometryReference(TextPos),
     #[error("Geometry '{0}' is referenced at line {1} but not found")]
